requirements.txt
asgiref==3.6.0
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==3.2
django-allauth==0.54.0
django-braces==1.15.0
django-oauth-toolkit==2.2.0
django-rest-auth==0.9.5
django-rest-framework-social-oauth2==1.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-spectacular==0.26.2
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
jwcrypto==1.4.2
oauthlib==3.2.2
psycopg2==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==5.4.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
wrapt==1.15.0
